// packages/sheets-drawing/src/commands/commands/interfaces.ts
/**
 * Copyright 2023-present DreamNum Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import type { DrawingTypeEnum } from '@univerjs/core';
import type { ISheetDrawing } from '@univerjs/sheets-drawing';

export interface IInsertDrawingCommandParams {
    unitId: string;
    drawings: ISheetDrawing[];
}

export interface IDeleteDrawingCommandParam {
    unitId: string;
    subUnitId: string;
    drawingId: string;
    drawingType: DrawingTypeEnum;
}

export interface IDeleteDrawingCommandParams {
    unitId: string;
    drawings: IDeleteDrawingCommandParam[];
}

export interface ISetDrawingCommandParams {
    unitId: string;
    drawings: Partial<ISheetDrawing>[];
}